MealListActivity 75% complete
also fixed cast error in Meal.java

+++ Project/Code/app/src/main/java/com/uottawa/seg2105/grp1/mealer/model/Meal.java
@@ -137,7 +137,7 @@ public final class Meal implements IRepositoryEntity {
         this.cuisine = (String) cuisine;
         this.ingredients = (String) ingredients;
         this.allergens = (String) allergens;
-        this.price = (int) price;
+        this.price = ((Long) price).intValue();
         this.description = (String) description;
         this.cook = cook;
         this.isOffered = (boolean) isOffered;
